add split color utilities
