+++ src/lib/color.ts
@@ -1,6 +1,25 @@
 import { parse, rgb, hsl, hwb, lab, lch, oklab, oklch } from "culori";
-import { round, floor, limiter } from "./utils";
-import { getLocalTheme } from "./theme";
+import { round, limiter } from "./utils";
+import {
+  formatHex,
+  formatRgb,
+  formatHsl,
+  formatHwb,
+  formatLab,
+  formatLch,
+  formatOklab,
+  formatOklch,
+} from "./format";
+import {
+  parseHex,
+  parseRgb,
+  parseHsl,
+  parseHwb,
+  parseLab,
+  parseLch,
+  parseOklab,
+  parseOklch,
+} from "~/lib/parse";
 import { createStore } from "zustand";
 
 export type ThemeVariant = "auto" | "light" | "dark";
@@ -88,77 +107,6 @@ export type ColorAction = {
 
 export type ColorStore = ColorState & ColorAction;
 
-export const formatHex = (newColor: RgbColor | RgbColorMode): string => {
-  const { r, g, b } = newColor;
-
-  const clamp = (value: number) => Math.max(0, Math.min(1, value || 0));
-  const fixup = (value: number) => Math.round(clamp(value) * 255);
-
-  const red = fixup(r);
-  const green = fixup(g);
-  const blue = fixup(b);
-
-  return (
-    "#" + ((1 << 24) | (red << 16) | (green << 8) | blue).toString(16).slice(1)
-  );
-};
-export const formatRgb = (newColor: RgbColor | RgbColorMode): string => {
-  const { r, g, b } = newColor;
-  const red = round(r * 255);
-  const green = round(g * 255);
-  const blue = round(b * 255);
-
-  return `rgb(${red} ${green} ${blue})`;
-};
-export const formatHsl = (newColor: HslColor | HslColorMode): string => {
-  const { h, s, l } = newColor;
-  const hue = round(h || 0, 2);
-  const saturation = round(s * 100, 2);
-  const lightness = round(l * 100, 2);
-
-  return `hsl(${hue} ${saturation}% ${lightness}%)`;
-};
-export const formatHwb = (newColor: HwbColor | HwbColorMode): string => {
-  const { h, w, b } = newColor;
-  const hue = round(h || 0, 2);
-  const whiteness = round(w * 100, 2);
-  const blackness = round(b * 100, 2);
-
-  return `hwb(${hue} ${whiteness}% ${blackness}%)`;
-};
-export const formatLab = (newColor: LabColor | LabColorMode): string => {
-  const { l, a, b } = newColor;
-  const lightness = round(l, 3);
-  const greenRed = round(a, 3);
-  const blueYellow = round(b, 3);
-
-  return `lab(${lightness} ${greenRed} ${blueYellow})`;
-};
-export const formatLch = (newColor: LchColor | LchColorMode): string => {
-  const { l, c, h } = newColor;
-  const lightness = round(l, 3);
-  const chroma = round(c, 3);
-  const hue = round(h || 0, 2);
-
-  return `lch(${lightness} ${chroma} ${hue})`;
-};
-export const formatOklab = (newColor: OklabColor | OklabColorMode): string => {
-  const { l, a, b } = newColor;
-  const lightness = round(l, 3);
-  const greenRed = round(a, 3);
-  const blueYellow = round(b, 3);
-
-  return `oklab(${lightness} ${greenRed} ${blueYellow})`;
-};
-export const formatOklch = (newColor: OklchColor | OklchColorMode): string => {
-  const { l, c, h } = newColor;
-  const lightness = round(l, 3);
-  const chroma = round(c, 3);
-  const hue = round(h || 0, 2);
-
-  return `oklch(${lightness} ${chroma} ${hue})`;
-};
-
 export const createColorStore = (initValue: ColorState) => {
   return createStore<ColorStore>()((set) => ({
     ...initValue,
@@ -244,85 +192,17 @@ export const createColorStore = (initValue: ColorState) => {
   }));
 };
 
-const colorHex = (newColor: string | AnyColorMode): string => {
-  const _rgb = rgb(newColor) || { r: 0, g: 0, b: 0 };
-  return formatHex(_rgb);
-};
-const colorRgb = (newColor: string): { color: RgbColor; css: string } => {
-  const _rgb = rgb(newColor) || { r: 0, g: 0, b: 0 };
-
-  return { color: { r: _rgb.r, g: _rgb.g, b: _rgb.b }, css: formatRgb(_rgb) };
-};
-const colorHsl = (newColor: string): { color: HslColor; css: string } => {
-  const _hsl = hsl(newColor) || { h: 0, s: 0, l: 0 };
-
+export const createColor = (byHex: string) => {
   return {
-    color: { h: _hsl.h || 0, s: _hsl.s, l: _hsl.l },
-    css: formatHsl(_hsl),
+    hex: byHex,
+    rgb: parseRgb(byHex),
+    hsl: parseHsl(byHex),
+    hwb: parseHwb(byHex),
+    lab: parseLab(byHex),
+    lch: parseLch(byHex),
+    oklab: parseOklab(byHex),
+    oklch: parseOklch(byHex),
   };
-};
-const colorHwb = (newColor: string): { color: HwbColor; css: string } => {
-  const _hwb = hwb(newColor) || { h: 0, w: 0, b: 0 };
-
-  return {
-    color: { h: _hwb.h || 0, w: _hwb.w, b: _hwb.b },
-    css: formatHwb(_hwb),
-  };
-};
-const colorLab = (newColor: string): { color: LabColor; css: string } => {
-  const _lab = lab(newColor) || { l: 0, a: 0, b: 0 };
-
-  return { color: { l: _lab.l, a: _lab.a, b: _lab.b }, css: formatLab(_lab) };
-};
-const colorLch = (newColor: string): { color: LchColor; css: string } => {
-  const _lch = lch(newColor) || { l: 0, c: 0, h: 0 };
-
-  return {
-    color: { l: _lch.l, c: _lch.c, h: _lch.h || 0 },
-    css: formatLch(_lch),
-  };
-};
-const colorOklab = (newColor: string): { color: OklabColor; css: string } => {
-  const _oklab = oklab(newColor) || { l: 0, a: 0, b: 0 };
-
-  return {
-    color: { l: _oklab.l, a: _oklab.a, b: _oklab.b },
-    css: formatOklab(_oklab),
-  };
-};
-const colorOklch = (newColor: string): { color: OklchColor; css: string } => {
-  const _oklch = oklch(newColor) || { l: 0, c: 0, h: 0 };
-
-  return {
-    color: { l: _oklch.l, c: _oklch.c, h: _oklch.h || 0 },
-    css: formatOklch(_oklch),
-  };
-};
-
-export const createColor = (newColor: string): ColorState => {
-  const _rgb = colorRgb(newColor);
-  const _hsl = colorHsl(newColor);
-  const _hwb = colorHwb(newColor);
-  const _lab = colorLab(newColor);
-  const _lch = colorLch(newColor);
-  const _oklab = colorOklab(newColor);
-  const _oklch = colorOklch(newColor);
-  const _theme = getLocalTheme();
-
-  const store: ColorState = {
-    hex: formatHex(_rgb.color),
-    rgb: _rgb,
-    hsl: _hsl,
-    hwb: _hwb,
-    lab: _lab,
-    lch: _lch,
-    oklab: _oklab,
-    oklch: _oklch,
-    mode: "oklch",
-    theme: _theme,
-  };
-
-  return store;
 };
 
 export const isValidColor = (newColor: string): boolean => {
@@ -361,161 +241,10 @@ export const randomColor = (): string => {
   const saturation = round(limiter(month * 12, 0.64, 0.96), 4);
   const lightness = round(limiter(day * 30, 0.32, 0.64), 4);
 
-  return colorHex({
+  return parseHex({
     mode: "hsl",
     h: hue,
     s: saturation,
     l: lightness,
   });
-};
-
-export const parseColor = () => {
-  return {
-    hex: colorHex,
-    rgb: colorRgb,
-    hsl: colorHsl,
-    hwb: colorHwb,
-    lab: colorLab,
-    lch: colorLch,
-    oklab: colorOklab,
-    oklch: colorOklch,
-  };
-};
-
-export const stringifyColor = (newColor: AnyColorMode): string => {
-  switch (newColor.mode) {
-    case "rgb":
-      return formatRgb(rgb(newColor));
-      break;
-    case "hsl":
-      return formatHsl(hsl(newColor));
-      break;
-    case "hwb":
-      return formatHwb(hwb(newColor));
-      break;
-    case "lab":
-      return formatLab(lab(newColor));
-      break;
-    case "lch":
-      return formatLch(lch(newColor));
-      break;
-    case "oklab":
-      return formatOklab(oklab(newColor));
-      break;
-    case "oklch":
-      return formatOklch(oklch(newColor));
-  }
-};
-
-interface ReadabilityColor {
-  level: "AA" | "AAA";
-  size: "normal" | "large";
-}
-
-export const getBrightness = (newColor: RgbColor): number => {
-  const { r, g, b } = newColor;
-
-  const red = round(r * 255);
-  const green = round(g * 255);
-  const blue = round(b * 255);
-
-  return (red * 299 + green * 587 + blue * 114) / 1000 / 255;
-};
-export const getLuminance = (newColor: RgbColor): number => {
-  const linear = (value: number) => {
-    const ratio = value / 255;
-    return ratio < 0.04045
-      ? ratio / 12.92
-      : Math.pow((ratio + 0.055) / 1.055, 2.4);
-  };
-
-  const { r, g, b } = newColor;
-
-  const red = linear(r * 255);
-  const green = linear(g * 255);
-  const blue = linear(b * 255);
-
-  return 0.2126 * red + 0.7152 * green + 0.0722 * blue;
-};
-export const getContrast = (
-  foreground: RgbColor,
-  background: RgbColor,
-): number => {
-  const fg = getLuminance(foreground);
-  const bg = getLuminance(background);
-
-  return fg > bg ? (fg + 0.05) / (bg + 0.05) : (bg + 0.05) / (fg + 0.05);
-};
-export const getMinimalContrast = ({
-  level = "AA",
-  size = "normal",
-}: ReadabilityColor): number => {
-  if (level === "AAA" && size === "normal") return 7;
-  if (level === "AA" && size === "large") return 3;
-  return 4.5;
-};
-
-export const brightness = (newColor: RgbColor): string => {
-  const color = round(getBrightness(newColor), 2);
-  const level = round(color * 100);
-
-  if (color >= 0.5) {
-    return `${level}% (Light)`;
-  } else {
-    return `${level}% (Dark)`;
-  }
-};
-export const luminance = (newColor: RgbColor): string => {
-  const color = round(getLuminance(newColor), 2);
-  const level = round(color * 100);
-
-  return `${level}%`;
-};
-export const contrast = (
-  foreground: RgbColor,
-  background: RgbColor,
-): string => {
-  const color = floor(getContrast(foreground, background), 2);
-
-  return `${color}:1`;
-};
-export const isReadable = (
-  foreground: RgbColor,
-  background: RgbColor,
-  options: ReadabilityColor,
-): boolean => {
-  const contrast = floor(getContrast(foreground, background), 2);
-  const readable = getMinimalContrast(options);
-
-  return contrast >= readable;
-};
-
-export const inspectColor = (foreground: RgbColor, background: RgbColor) => {
-  const average = contrast(foreground, background);
-  const check = (readable: boolean) => (readable ? "Pass" : "Fail");
-
-  const normalAA = check(
-    isReadable(foreground, background, { level: "AA", size: "normal" }),
-  );
-  const normalAAA = check(
-    isReadable(foreground, background, { level: "AAA", size: "normal" }),
-  );
-  const largeAA = check(
-    isReadable(foreground, background, { level: "AA", size: "large" }),
-  );
-  const largeAAA = check(
-    isReadable(foreground, background, { level: "AAA", size: "large" }),
-  );
-
-  return {
-    ratio: average,
-    normal: {
-      aa: normalAA,
-      aaa: normalAAA,
-    },
-    large: {
-      aa: largeAA,
-      aaa: largeAAA,
-    },
-  };
 };
