+++ src/lib/format.ts
@@ -1,0 +1,89 @@
+import type {
+  RgbColor,
+  HslColor,
+  HwbColor,
+  LabColor,
+  LchColor,
+  OklabColor,
+  OklchColor,
+} from "./color";
+
+import { round } from "./utils";
+
+export const formatHex = (newColor: RgbColor): string => {
+  const { r, g, b } = newColor;
+
+  const clamp = (value: number) => Math.max(0, Math.min(1, value || 0));
+  const fixup = (value: number) => Math.round(clamp(value) * 255);
+
+  const red = fixup(r);
+  const green = fixup(g);
+  const blue = fixup(b);
+
+  return (
+    "#" + ((1 << 24) | (red << 16) | (green << 8) | blue).toString(16).slice(1)
+  );
+};
+
+export const formatRgb = (newColor: RgbColor): string => {
+  const { r, g, b } = newColor;
+  const red = round(r * 255);
+  const green = round(g * 255);
+  const blue = round(b * 255);
+
+  return `rgb(${red} ${green} ${blue})`;
+};
+
+export const formatHsl = (newColor: HslColor): string => {
+  const { h, s, l } = newColor;
+  const hue = round(h || 0, 2);
+  const saturation = round(s * 100, 2);
+  const lightness = round(l * 100, 2);
+
+  return `hsl(${hue} ${saturation}% ${lightness}%)`;
+};
+
+export const formatHwb = (newColor: HwbColor): string => {
+  const { h, w, b } = newColor;
+  const hue = round(h || 0, 2);
+  const whiteness = round(w * 100, 2);
+  const blackness = round(b * 100, 2);
+
+  return `hwb(${hue} ${whiteness}% ${blackness}%)`;
+};
+
+export const formatLab = (newColor: LabColor): string => {
+  const { l, a, b } = newColor;
+  const lightness = round(l, 3);
+  const greenRed = round(a, 3);
+  const blueYellow = round(b, 3);
+
+  return `lab(${lightness} ${greenRed} ${blueYellow})`;
+};
+
+export const formatLch = (newColor: LchColor): string => {
+  const { l, c, h } = newColor;
+  const lightness = round(l, 3);
+  const chroma = round(c, 3);
+  const hue = round(h || 0, 2);
+
+  return `lch(${lightness} ${chroma} ${hue})`;
+};
+
+export const formatOklab = (newColor: OklabColor): string => {
+  const { l, a, b } = newColor;
+  const lightness = round(l, 3);
+  const greenRed = round(a, 3);
+  const blueYellow = round(b, 3);
+
+  return `oklab(${lightness} ${greenRed} ${blueYellow})`;
+};
+
+export const formatOklch = (newColor: OklchColor): string => {
+  const { l, c, h } = newColor;
+  const lightness = round(l, 3);
+  const chroma = round(c, 3);
+  const hue = round(h || 0, 2);
+
+  return `oklch(${lightness} ${chroma} ${hue})`;
+};
